+++ crates/huub/src/solver/engine.rs
@@ -937,7 +937,7 @@ impl IntInspectionActions<State> for IntVarRef {
 		ctx.int_vars[*self].get_upper_bound(&ctx.trail)
 	}
 
-	fn check_int_in_domain(&self, ctx: &State, val: IntVal) -> bool {
+	fn check_in_domain(&self, ctx: &State, val: IntVal) -> bool {
 		let (lb, ub) = self.get_bounds(ctx);
 		if lb <= val && val <= ub {
 			let eq_lit = self.try_lit(ctx, IntLitMeaning::Eq(val));
@@ -975,241 +975,9 @@ impl IntInspectionActions<State> for IntVarRef {
 	}
 }
 
-impl IntExplanationActions<State> for IntView {
-	fn get_lit_relaxed(
-		&self,
-		ctx: &State,
-		mut meaning: IntLitMeaning,
-	) -> (BoolView, IntLitMeaning) {
-		debug_assert!(
-			!matches!(meaning, IntLitMeaning::Eq(_)),
-			"relaxed integer literals are not yet supported for IntLitMeaning::Eq(_)"
-		);
-		// Transform literal meaning if view is a linear transformation
-		if let IntViewInner::Linear { transformer, .. } | IntViewInner::Bool { transformer, .. } =
-			self.0
-		{
-			match transformer.rev_transform_lit(meaning) {
-				Ok(m) => meaning = m,
-				Err(v) => return (BoolView(BoolViewInner::Const(v)), meaning),
-			}
-		}
-
-		// Get the boolean view that is currently `true` and implies the requested
-		// `meaning`, as well as the actual (possibly relaxed) meaning that is
-		// represented.
-		let (bv, meaning) = match self.0 {
-			IntViewInner::VarRef(iv) | IntViewInner::Linear { var: iv, .. } => {
-				iv.get_lit_relaxed(ctx, meaning)
-			}
-			IntViewInner::Const(c) => (
-				BoolView(BoolViewInner::Const(match meaning {
-					IntLitMeaning::GreaterEq(i) => c >= i,
-					IntLitMeaning::Less(i) => c < i,
-					IntLitMeaning::Eq(i) => c == i,
-					IntLitMeaning::NotEq(i) => c != i,
-				})),
-				meaning,
-			),
-			IntViewInner::Bool { lit, .. } => {
-				let (b_meaning, negated) =
-					if matches!(meaning, IntLitMeaning::NotEq(_) | IntLitMeaning::Less(_)) {
-						(!meaning, true)
-					} else {
-						(meaning, false)
-					};
-				let bv = BoolView(match b_meaning {
-					IntLitMeaning::GreaterEq(1) => BoolViewInner::Lit(lit),
-					IntLitMeaning::GreaterEq(i) if i > 1 => BoolViewInner::Const(false),
-					IntLitMeaning::GreaterEq(_) => BoolViewInner::Const(true),
-					IntLitMeaning::Eq(0) => BoolViewInner::Lit(!lit),
-					IntLitMeaning::Eq(1) => BoolViewInner::Lit(lit),
-					IntLitMeaning::Eq(_) => BoolViewInner::Const(false),
-					_ => unreachable!(),
-				});
-				(if negated { !bv } else { bv }, meaning)
-			}
-		};
-
-		// Transform the meaning back to fit the original view if it was linearly
-		// transformed
-		let meaning = if let IntViewInner::Linear { transformer, .. }
-		| IntViewInner::Bool { transformer, .. } = self.0
-		{
-			transformer.transform_lit(meaning)
-		} else {
-			meaning
-		};
-		(bv, meaning)
-	}
-}
-
 impl BoolInspectionActions<State> for RawLit {
 	fn get_val(&self, ctx: &State) -> Option<bool> {
 		ctx.trail.get_sat_value(*self)
-	}
-}
-
-impl IntInspectionActions<State> for IntView {
-	type Atom = BoolView;
-
-	fn get_lit_meaning(&self, ctx: &State, lit: Self::Atom) -> Option<IntLitMeaning> {
-		match self.0 {
-			IntViewInner::VarRef(iv) | IntViewInner::Linear { var: iv, .. } => {
-				let mut meaning = iv.get_lit_meaning(ctx, lit)?;
-				if let IntViewInner::Linear { transformer, .. } = self.0 {
-					meaning = transformer.transform_lit(meaning);
-				}
-				Some(meaning)
-			}
-			IntViewInner::Const(_) => None,
-			IntViewInner::Bool {
-				lit: var_lit,
-				transformer,
-			} => {
-				let BoolViewInner::Lit(lit) = lit.0 else {
-					return None;
-				};
-				if lit.var() != var_lit.var() {
-					return None;
-				}
-				let mut meaning = IntLitMeaning::GreaterEq(1);
-				if var_lit != lit {
-					meaning = !meaning;
-				}
-				meaning = transformer.transform_lit(meaning);
-				Some(meaning)
-			}
-		}
-	}
-
-	fn get_lower_bound(&self, ctx: &State) -> IntVal {
-		match self.0 {
-			IntViewInner::VarRef(var) => var.get_lower_bound(ctx),
-			IntViewInner::Const(c) => c,
-			IntViewInner::Linear { transformer, var } => {
-				transformer.transform(if transformer.positive_scale() {
-					var.get_lower_bound(ctx)
-				} else {
-					var.get_upper_bound(ctx)
-				})
-			}
-			IntViewInner::Bool { transformer, lit } => transformer
-				.transform(lit.get_val(ctx).unwrap_or(!transformer.positive_scale()) as IntVal),
-		}
-	}
-
-	fn get_upper_bound(&self, ctx: &State) -> IntVal {
-		match self.0 {
-			IntViewInner::VarRef(var) => var.get_upper_bound(ctx),
-			IntViewInner::Const(c) => c,
-			IntViewInner::Linear { transformer, var } => {
-				transformer.transform(if transformer.positive_scale() {
-					var.get_upper_bound(ctx)
-				} else {
-					var.get_lower_bound(ctx)
-				})
-			}
-			IntViewInner::Bool { transformer, lit } => transformer
-				.transform(lit.get_val(ctx).unwrap_or(transformer.positive_scale()) as IntVal),
-		}
-	}
-
-	fn check_int_in_domain(&self, ctx: &State, val: IntVal) -> bool {
-		let (lb, ub) = self.get_bounds(ctx);
-		if lb <= val && val <= ub {
-			let eq_lit = self.try_lit(ctx, IntLitMeaning::Eq(val));
-			if let Some(eq_lit) = eq_lit {
-				eq_lit.get_val(ctx).unwrap_or(true)
-			} else {
-				true
-			}
-		} else {
-			false
-		}
-	}
-
-	fn get_lower_bound_lit(&self, ctx: &State) -> BoolView {
-		match self.0 {
-			IntViewInner::VarRef(var) => var.get_lower_bound_lit(ctx),
-			IntViewInner::Linear { transformer, var } => {
-				if transformer.positive_scale() {
-					var.get_lower_bound_lit(ctx)
-				} else {
-					var.get_upper_bound_lit(ctx)
-				}
-			}
-			IntViewInner::Const(_) => BoolView(BoolViewInner::Const(true)),
-			IntViewInner::Bool { lit, transformer } => {
-				BoolView(match (lit.get_val(ctx), transformer.positive_scale()) {
-					(Some(true), true) => BoolViewInner::Lit(lit),
-					(Some(false), false) => BoolViewInner::Lit(!lit),
-					_ => BoolViewInner::Const(true),
-				})
-			}
-		}
-	}
-
-	fn get_upper_bound_lit(&self, ctx: &State) -> BoolView {
-		match self.0 {
-			IntViewInner::VarRef(var) => var.get_upper_bound_lit(ctx),
-			IntViewInner::Linear { transformer, var } => {
-				if transformer.positive_scale() {
-					var.get_upper_bound_lit(ctx)
-				} else {
-					var.get_lower_bound_lit(ctx)
-				}
-			}
-			IntViewInner::Const(_) => BoolView(BoolViewInner::Const(true)),
-			IntViewInner::Bool { lit, transformer } => {
-				BoolView(match (lit.get_val(ctx), transformer.positive_scale()) {
-					(Some(false), true) => BoolViewInner::Lit(!lit),
-					(Some(true), false) => BoolViewInner::Lit(lit),
-					_ => BoolViewInner::Const(true),
-				})
-			}
-		}
-	}
-
-	fn try_lit(&self, ctx: &State, mut meaning: IntLitMeaning) -> Option<BoolView> {
-		if let IntViewInner::Linear { transformer, .. } | IntViewInner::Bool { transformer, .. } =
-			self.0
-		{
-			match transformer.rev_transform_lit(meaning) {
-				Ok(m) => meaning = m,
-				Err(v) => return Some(BoolView(BoolViewInner::Const(v))),
-			}
-		}
-
-		match self.0 {
-			IntViewInner::VarRef(var) | IntViewInner::Linear { var, .. } => {
-				var.try_lit(ctx, meaning)
-			}
-			IntViewInner::Const(c) => Some(BoolView(BoolViewInner::Const(match meaning {
-				IntLitMeaning::Eq(i) => c == i,
-				IntLitMeaning::NotEq(i) => c != i,
-				IntLitMeaning::GreaterEq(i) => c >= i,
-				IntLitMeaning::Less(i) => c < i,
-			}))),
-			IntViewInner::Bool { lit, .. } => {
-				let (meaning, negated) =
-					if matches!(meaning, IntLitMeaning::NotEq(_) | IntLitMeaning::Less(_)) {
-						(!meaning, true)
-					} else {
-						(meaning, false)
-					};
-				let bv = BoolView(match meaning {
-					IntLitMeaning::Eq(0) => BoolViewInner::Lit(!lit),
-					IntLitMeaning::Eq(1) => BoolViewInner::Lit(lit),
-					IntLitMeaning::Eq(_) => BoolViewInner::Const(false),
-					IntLitMeaning::GreaterEq(1) => BoolViewInner::Lit(lit),
-					IntLitMeaning::GreaterEq(i) if i > 1 => BoolViewInner::Const(false),
-					IntLitMeaning::GreaterEq(_) => BoolViewInner::Const(true),
-					_ => unreachable!(),
-				});
-				Some(if negated { !bv } else { bv })
-			}
-		}
 	}
 }
